move generator to pkg/util/kslib
